Create a tag-like ref of current HEAD
git describe --tags --always, allows us to get a unique version
describing the current HEAD of the git resource. This has a nice
semver format that can be used in a Concourse pipeline for
versioning artifacts and releases.

Falls back to a short commit SHA-1 ref if the repo was never tagged.

Signed-off-by: Salah-Eddine Missri <salah.missri@pix4d.com>